Replace karma-coverage with karma-js-coverage, remove CoffeeScript deps
karma-coverage hasn't been released on npm, and uses CoffeeScript as a git
submodule which results in a number of issues (see below). Since we don't need
CoffeeScript support, we should be able to use karma-js-coverage from npm in its
place. This reduces dependencies, and should fix the nix packaging issues caused
by the frequently changing sha256 of karma-coverage.

https://github.com/karma-runner/karma-coverage/issues/60
https://github.com/michaelficarra/CoffeeScriptRedux/issues/60